move internal/value to pkg/value
