+++ src/pkg/value/testdata/schema/invalid-range.yaml
