+++ src/pkg/value/testdata/schema/invalid-values.yaml
