+++ src/pkg/value/testdata/schema/simple.schema.json
