+++ src/pkg/value/testdata/schema/valid-values.yaml
