+++ src/pkg/value/testdata/valid/arrays.yaml
