+++ src/pkg/value/testdata/valid/complex.yaml
