+++ src/pkg/value/testdata/valid/empty.yaml
