+++ src/pkg/value/testdata/valid/merge-overwrite.yaml
