+++ src/pkg/value/testdata/valid/merge1.yaml
