+++ src/pkg/value/testdata/valid/merge2.yaml
