+++ src/pkg/value/testdata/valid/simple.yaml
